Split Code into Multiple Scripts
Hoping to avoid having one massive file. There are still a lot of bugs to squash.
