+++ src/img/colors.py
@@ -1,0 +1,5 @@
+WHITE = (255, 255, 255)
+BLACK = (0, 0, 0)
+GREY = (128, 128, 128)
+RED = (255, 0, 0)
+BLUE = (0, 0, 255)
