+++ src/properties.py
@@ -1,0 +1,5 @@
+class Properties:  # Tile and Grid Size
+    TILE_MAP = []
+    TILE_DIM = 16
+    GRID_WIDTH = 0
+    GRID_HEIGHT = 0
