+++ src/tiles/red.py
@@ -1,19 +1,16 @@
-import main
+from src.tiles.tile import Tile
+from src.img import colors as c
 
 
-RED = (255, 0, 0)
-BLUE = (0, 0, 255)
-
-
-class RedBlue(main.Tile):
+class RedBlue(Tile):
     def __init__(self):
         self._clicked = False
-        super().__init__(BLUE)
+        super().__init__(c.BLUE)
 
     def click(self):
         print("Clicked on a RedBlue tile")
         self._clicked = not self._clicked
         if self._clicked:
-            self._color = RED
+            self._color = c.RED
         else:
-            self._color = BLUE
+            self._color = c.BLUE
